requirements.txt
# Core
Pillow
matplotlib
opencv-python
numpy
jupyter

# Panorama OCR (Chapter 02)
# For macOS: pip install "panoocr[macocr]"
# For Windows/Linux: pip install "panoocr[paddleocr]"
panoocr[macocr]

# Google Street View (Chapter 04)
streetlevel

# Object Segmentation (Chapter 03)
# Requires GPU for reasonable performance and HuggingFace auth.
# pip install "panosam[full]"
panosam[full]
